Tests/LibWeb/Text/expected/Messaging/Messaging-post-channel-over-channel.txt

Port1: "Hello"
Port1: {"foo":{}}
Port2: "Hello"
Port3: "Hello from the transferred port"
Port1: "DONE"
Port2: "DONE"
